LUCENE-4456: add a test that MDW actually detects and fails on unref'd files
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/lucene4456@1392861 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/test-framework/src/java/org/apache/lucene/store/MockDirectoryWrapper.java
@@ -214,7 +214,7 @@ public class MockDirectoryWrapper extends BaseDirectoryWrapper {
       String name = it.next();
       int damage = randomState.nextInt(5);
       String action = null;
-      if (damage == 0) {
+      if (damage == 0 && !name.equals("TEST_THE_TESTER_FILE_NEVER_DELETED_ON_CRASH")) {
         action = "deleted";
         deleteFile(name, true);
       } else if (damage == 1) {
